requirements-gemma4-lora.txt
sentencepiece
protobuf
datasets==4.3.0
huggingface_hub>=0.34.0
hf_transfer
unsloth
unsloth_zoo
bitsandbytes
accelerate
xformers==0.0.34
peft
trl
triton
torchao>=0.16.0
transformers==5.5.0
timm
torchcodec
